chore: Drone Try caching (#42)
This uses the official Drone S3 Caching plugin

It works pretty well, but it doesn't do caching exactly how I would like. One of the big issues is that it only caches by branch, without support for lockfiles, or arbitrary (runtime decided) cache keys.

So I think this will work for now, but in the future I think I want to write my own plugin to replace this. But for now I think this will be better than nothing!

* Try caching

* Remove debug and set cache to flush 30 days

* Sign drone file

* Do flush and rebuild seperately

* Fix name collision